requirements.txt
# For MQTT
paho-mqtt==1.6.1

# For databases
sqlalchemy==1.4.46
pymongo==4.3.3
neo4j==5.5.0

# For data processing
pandas==1.5.3
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2

# For natural language processing
textblob==0.17.1
nltk==3.8.1
scikit-learn==1.2.2

# For social media APIs
tweepy==4.12.1
praw==7.7.0

# For scheduled tasks
schedule==1.1.0

# For environment variables
python-dotenv==1.0.0

# General utility tools
click==8.1.3
tqdm==4.65.0

# For web interface (optional)
flask==2.2.3
dash==2.9.2
plotly==5.14.1
